requirements.txt
# For running the GUI (stdlib only - tkinter is built-in)
# No dependencies required to run mewgenics_backup_gui.py

# For building Windows executable (install on Windows)
pyinstaller>=6.0.0

# For converting PNG to ICO (icons/png_to_ico.py)
Pillow>=10.0.0
